feat: 카테고리 재정렬 응답 객체 병합
- CategorySequenceResponse를 CategoryReorderResponse 배부로 이동
