+++ src/main/java/courseitda/workspace/ui/dto/response/CategoryReorderResponse.java
@@ -2,6 +2,7 @@ package courseitda.workspace.ui.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import courseitda.workspace.application.dto.response.UpdateCategorySequenceResult;
+import courseitda.workspace.application.dto.response.UpdateCategorySequenceResult.CategorySequenceResult;
 import java.util.List;
 
 public record CategoryReorderResponse(
@@ -14,5 +15,18 @@ public record CategoryReorderResponse(
                         .map(CategorySequenceResponse::from)
                         .toList()
         );
+    }
+
+    public record CategorySequenceResponse(
+            Long id,
+            Integer sequence
+    ) {
+
+        public static CategorySequenceResponse from(final CategorySequenceResult result) {
+            return new CategorySequenceResponse(
+                    result.id(),
+                    result.sequence()
+            );
+        }
     }
 }
