--- src/main/java/courseitda/workspace/ui/dto/response/CategorySequenceResponse.java
@@ -1,16 +1,0 @@
-package courseitda.workspace.ui.dto.response;
-
-import courseitda.workspace.application.dto.response.UpdateCategorySequenceResult.CategorySequenceResult;
-
-public record CategorySequenceResponse(
-        Long id,
-        Integer sequence
-) {
-
-    public static CategorySequenceResponse from(final CategorySequenceResult result) {
-        return new CategorySequenceResponse(
-                result.id(),
-                result.sequence()
-        );
-    }
-}
